requirements.txt
python-telegram-bot[webhooks]==20.6
gTTS==2.5.1
deep-translator==1.11.4
